splinting now gets up to scicos_blocks module

+++ scilab/modules/graphics/sci_gateway/c/sci_xclick.c
@@ -31,102 +31,102 @@
 #include "getGraphicObjectProperty.h"
 
 /*--------------------------------------------------------------------------*/
-int sci_xclick(char *fname,unsigned long fname_len)
+int sci_xclick(char *fname, unsigned long fname_len)
 {
-  int one = 1, three = 3, rep = 0;
-  int istr = 0;
-  //int iflag = 0;
+    int one = 1, three = 3, rep = 0;
+    int istr = 0;
+    //int iflag = 0;
 
-  int mouseButtonNumber = 0;
-  char * menuCallback = NULL;
-  char *pstWindowUID = NULL;
-  int pixelCoords[2];
-  double userCoords2D[2];
+    int mouseButtonNumber = 0;
+    char * menuCallback = NULL;
+    char *pstWindowUID = NULL;
+    int pixelCoords[2];
+    double userCoords2D[2];
 
-  int iFigureId = 0;
-  int *piFigureId = &iFigureId;
+    int iFigureId = 0;
+    int *piFigureId = &iFigureId;
 
-  CheckRhs(-1, 1) ;
-  CheckLhs(1, 5) ;
+    CheckRhs(-1, 1) ;
+    CheckLhs(1, 5) ;
 
-  //iflag = ( Rhs >= 1) ? 1 :0;
+    //iflag = ( Rhs >= 1) ? 1 :0;
 
-  // Select current figure or create it
-  getOrCreateDefaultSubwin();
+    // Select current figure or create it
+    getOrCreateDefaultSubwin();
 
-  // Call Java xclick
-  CallJxclick();
+    // Call Java xclick
+    CallJxclick();
 
-  // Get return values
-  mouseButtonNumber = getJxclickMouseButtonNumber();
-  pixelCoords[0] = (int) getJxclickXCoordinate();
-  pixelCoords[1] = (int) getJxclickYCoordinate();
-  pstWindowUID = getJxclickWindowID();
-  menuCallback = getJxclickMenuCallback();
+    // Get return values
+    mouseButtonNumber = getJxclickMouseButtonNumber();
+    pixelCoords[0] = (int) getJxclickXCoordinate();
+    pixelCoords[1] = (int) getJxclickYCoordinate();
+    pstWindowUID = getJxclickWindowID();
+    menuCallback = getJxclickMenuCallback();
 
-  // Convert pixel coordinates to user coordinates
-  // Conversion is not done if the user clicked on a menu (pixelCoords[*] == -1)
-  if (pixelCoords[0] != -1 && pixelCoords[1] != -1)
-  {
-    char* clickedSubwinUID = (char*)getCurrentSubWin();
-    sciGet2dViewCoordFromPixel(clickedSubwinUID, pixelCoords, userCoords2D);
-  }
-  else
-  {
-    userCoords2D[0] = pixelCoords[0];
-    userCoords2D[1] = pixelCoords[1];
-  }
+    // Convert pixel coordinates to user coordinates
+    // Conversion is not done if the user clicked on a menu (pixelCoords[*] == -1)
+    if (pixelCoords[0] != -1 && pixelCoords[1] != -1)
+    {
+        char* clickedSubwinUID = (char*)getCurrentSubWin();
+        sciGet2dViewCoordFromPixel(clickedSubwinUID, pixelCoords, userCoords2D);
+    }
+    else
+    {
+        userCoords2D[0] = (double)pixelCoords[0];
+        userCoords2D[1] = (double)pixelCoords[1];
+    }
 
-  if (Lhs == 1)
-  {
-    LhsVar(1) = Rhs+1;
-    CreateVar(Rhs+1,MATRIX_OF_DOUBLE_DATATYPE,&one,&three,&rep);
-    *stk(rep) = (double) mouseButtonNumber;
-    *stk(rep + 1) = userCoords2D[0];
-    *stk(rep + 2) = userCoords2D[1];
-  }
-  else
-  {
-    LhsVar(1) = Rhs+1;
-    CreateVar(Rhs+1,MATRIX_OF_DOUBLE_DATATYPE,&one,&one,&rep);
-    *stk(rep) = (double) mouseButtonNumber;
-  }
+    if (Lhs == 1)
+    {
+        LhsVar(1) = Rhs + 1;
+        CreateVar(Rhs + 1, MATRIX_OF_DOUBLE_DATATYPE, &one, &three, &rep);
+        *stk(rep) = (double) mouseButtonNumber;
+        *stk(rep + 1) = userCoords2D[0];
+        *stk(rep + 2) = userCoords2D[1];
+    }
+    else
+    {
+        LhsVar(1) = Rhs + 1;
+        CreateVar(Rhs + 1, MATRIX_OF_DOUBLE_DATATYPE, &one, &one, &rep);
+        *stk(rep) = (double) mouseButtonNumber;
+    }
 
-  if (Lhs >= 2)
-  {
-    LhsVar(2) = Rhs+2;
-    CreateVar(Rhs+2,MATRIX_OF_DOUBLE_DATATYPE,&one,&one,&rep);
-    *stk(rep) = userCoords2D[0];
-  }
+    if (Lhs >= 2)
+    {
+        LhsVar(2) = Rhs + 2;
+        CreateVar(Rhs + 2, MATRIX_OF_DOUBLE_DATATYPE, &one, &one, &rep);
+        *stk(rep) = userCoords2D[0];
+    }
 
-  if (Lhs >= 3)
-  {
-    LhsVar(3) = Rhs+3;
-    CreateVar(Rhs+3,MATRIX_OF_DOUBLE_DATATYPE,&one,&one,&rep);
-    *stk(rep) = userCoords2D[1];
-  }
+    if (Lhs >= 3)
+    {
+        LhsVar(3) = Rhs + 3;
+        CreateVar(Rhs + 3, MATRIX_OF_DOUBLE_DATATYPE, &one, &one, &rep);
+        *stk(rep) = userCoords2D[1];
+    }
 
-  if (Lhs >=4)
-  {
-    LhsVar(4) = Rhs+4;
-    CreateVar(Rhs+4,MATRIX_OF_DOUBLE_DATATYPE,&one,&one,&rep);
-    getGraphicObjectProperty(pstWindowUID, __GO_ID__, jni_int, (void**)&piFigureId);
-    *stk(rep) = (double) iFigureId;
-  }
+    if (Lhs >= 4)
+    {
+        LhsVar(4) = Rhs + 4;
+        CreateVar(Rhs + 4, MATRIX_OF_DOUBLE_DATATYPE, &one, &one, &rep);
+        getGraphicObjectProperty(pstWindowUID, __GO_ID__, jni_int, (void**)&piFigureId);
+        *stk(rep) = (double) iFigureId;
+    }
 
-  if (Lhs >= 5)
-  {
-    LhsVar(5) = Rhs+5;
-    istr = (int)strlen(menuCallback);
-    CreateVar(Rhs+5,STRING_DATATYPE,&istr,&one,&rep);
-    strncpy(cstk(rep),menuCallback,istr);
-  }
+    if (Lhs >= 5)
+    {
+        LhsVar(5) = Rhs + 5;
+        istr = (int)strlen(menuCallback);
+        CreateVar(Rhs + 5, STRING_DATATYPE, &istr, &one, &rep);
+        strncpy(cstk(rep), menuCallback, istr);
+    }
 
-  deleteJxclickString(menuCallback);
-  deleteJxclickString(pstWindowUID);
+    deleteJxclickString(menuCallback);
+    deleteJxclickString(pstWindowUID);
 
-  PutLhsVar();
+    PutLhsVar();
 
-  return 0;
+    return 0;
 }
 /*--------------------------------------------------------------------------*/
