+++ scilab/modules/graphics/src/c/getHandleProperty/get_callback_type_property.c
@@ -37,10 +37,10 @@ int get_callback_type_property(void* _pvCtx, char* pobjUID)
 
     if (piCallbackType == NULL)
     {
-        Scierror(999, _("'%s' property does not exist for this handle.\n"),"callback_type");
+        Scierror(999, _("'%s' property does not exist for this handle.\n"), "callback_type");
         return FALSE;
     }
 
-    return sciReturnDouble(_pvCtx, iCallbackType);
+    return sciReturnDouble(_pvCtx, (double)iCallbackType);
 }
 /*------------------------------------------------------------------------*/
