+++ scilab/modules/graphics/src/c/getHandleProperty/get_foreground_property.c
@@ -40,10 +40,10 @@ int get_foreground_property(void* _pvCtx, char* pobjUID)
 
     if (piLineColor == NULL)
     {
-        Scierror(999, _("'%s' property does not exist for this handle.\n"),"foreground");
+        Scierror(999, _("'%s' property does not exist for this handle.\n"), "foreground");
         return -1;
     }
 
-    return sciReturnDouble(_pvCtx, iLineColor);
+    return sciReturnDouble(_pvCtx, (double)iLineColor);
 }
 /*------------------------------------------------------------------------*/
