+++ scilab/modules/graphics/src/c/getHandleProperty/get_mark_foreground_property.c
@@ -41,10 +41,10 @@ int get_mark_foreground_property(void* _pvCtx, char* pobjUID)
 
     if (piMarkForeground == NULL)
     {
-        Scierror(999, _("'%s' property does not exist for this handle.\n"),"mark_foreground");
+        Scierror(999, _("'%s' property does not exist for this handle.\n"), "mark_foreground");
         return -1;
     }
 
-    return sciReturnDouble(_pvCtx, iMarkForeground);
+    return sciReturnDouble(_pvCtx, (double)iMarkForeground);
 }
 /*------------------------------------------------------------------------*/
