+++ scilab/modules/graphics/src/c/getHandleProperty/get_mark_size_property.c
@@ -39,10 +39,10 @@ int get_mark_size_property(void* _pvCtx, char* pobjUID)
 
     if (piMarkSize == NULL)
     {
-        Scierror(999, _("'%s' property does not exist for this handle.\n"),"mark_size");
+        Scierror(999, _("'%s' property does not exist for this handle.\n"), "mark_size");
         return -1;
     }
 
-    return sciReturnDouble(_pvCtx, iMarkSize);
+    return sciReturnDouble(_pvCtx, (double)iMarkSize);
 }
 /*------------------------------------------------------------------------*/
